+++ clients/include/lapack_host_reference.hpp
@@ -23,6 +23,20 @@ void cblas_gemm(hipsolverOperation_t transA,
                 int                  ldc);
 
 template <typename T>
+void cblas_ormqr_unmqr(hipsolverSideMode_t  side,
+                       hipsolverOperation_t trans,
+                       int                  m,
+                       int                  n,
+                       int                  k,
+                       T*                   A,
+                       int                  lda,
+                       T*                   Ipiv,
+                       T*                   C,
+                       int                  ldc,
+                       T*                   work,
+                       int                  sizeW);
+
+template <typename T>
 void cblas_geqrf(int m, int n, T* A, int lda, T* ipiv, T* work, int sizeW);
 
 template <typename T, typename W>